Task status is now updated correctly.

+++ src/main/java/uk/gov/hmcts/reform/dev/repositories/TaskRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.reform.dev.models.Task;
 
 @Repository
-public interface TaskRepository extends CrudRepository<Task, Integer>{}
+public interface TaskRepository extends CrudRepository<Task, Long>{}
